refactor(addNumberOfUsers): 사용자 수 +1 로직 util 폴더로 분리
